+++ src/lib/options/index.ts
@@ -1,3 +1,4 @@
 export * from './options.actions';
 export * from './options.models';
 export * from './options.reducer';
+export * from './validate-options';