+++ src/lib/renderer/render-initial-view.ts
@@ -2,7 +2,7 @@ import * as d3 from '../d3';
 
 import type { Store } from '../store';
 import type { Data } from '../data';
-import { getDateSlice, getText, getColor, safeName, getIconID } from '../utils';
+import { getDateSlice, getText, getColor, safeName, getIconID, countDecimals } from '../utils';
 import type { RenderOptions } from './render-options';
 import { calculateDimensions } from './calculate-dimensions';
 import { renderHeader } from './render-header';
@@ -20,6 +20,7 @@ export function renderInitialView(data: Data[], store: Store, renderOptions: Ren
   const CompleteDateSlice = getDateSlice(currentDate, data, store);
   const dateSlice = CompleteDateSlice.slice(0, topN);
   const lastDateIndex = dates.indexOf(currentDate) > 0 ? dates.indexOf(currentDate) - 1 : 0;
+  const valueDecimals = store.getState().options.valueDecimals;
   renderOptions.lastDate = dates[lastDateIndex];
 
   if (!root || dateSlice.length === 0) return;
@@ -97,7 +98,8 @@ export function renderInitialView(data: Data[], store: Store, renderOptions: Ren
       .enter()
       .append('text')
       .attr('class', 'label')
-      .classed('outside-bars', labelsPosition !== 'inside')
+      .classed('outside-bars', labelsPosition === 'outside')
+      .classed('hidden', labelsPosition === 'none')
       .attr('x', labelX)
       .attr('y', (d: Data) => barY(d) + barHalfHeight)
       .style('text-anchor', 'end')
@@ -114,7 +116,11 @@ export function renderInitialView(data: Data[], store: Store, renderOptions: Ren
       .attr('class', 'valueLabel')
       .attr('x', (d: Data) => x(d.value) + 5)
       .attr('y', (d: Data) => barY(d) + barHalfHeight)
-      .text((d: Data) => d3.format(',.0f')(d.lastValue as number));
+      .text((d: Data) =>
+        valueDecimals === 'preserve'
+          ? d.lastValue
+          : d3.format(`,.${countDecimals(d.lastValue ?? d.value)}f`)(d.lastValue as number),
+      );
 
     if (showIcons) {
       const defs = (renderOptions.defs = svg.append('svg:defs'));
@@ -136,6 +142,15 @@ export function renderInitialView(data: Data[], store: Store, renderOptions: Ren
         .attr('y', 0);
 
       svg
+        .append('clipPath')
+        .attr('id', 'clipPath-' + store.getState().container.element.id)
+        .append('rect')
+        .attr('x', x(0) + 1)
+        .attr('y', margin.top)
+        .attr('width', width)
+        .attr('height', height);
+
+      svg
         .selectAll('circle')
         .data(dateSlice, (d: Data) => d.name)
         .enter()
@@ -144,7 +159,8 @@ export function renderInitialView(data: Data[], store: Store, renderOptions: Ren
         .attr('cy', (d: Data) => y(d.rank as number) + barHalfHeight)
         .attr('r', iconSize / 2)
         .style('fill', 'transparent')
-        .style('fill', (d: Data) => `url(#${getIconID(d)})`);
+        .style('fill', (d: Data) => `url(#${getIconID(d)})`)
+        .attr('clip-path', `url(#clipPath-${store.getState().container.element.id})`);
     }
 
     const endY = height - margin.bottom;